yc-strategy-agent/requirements.txt
qdrant-client[fastembed]>=1.14.2
mistralai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0

